requirements.txt
Django>=5.1.7
djangorestframework
psycopg2-binary
python-decouple